+++ benchmark/bin/write.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use
+
 import 'dart:async';
 
 import 'package:objectbox_benchmark/benchmark.dart';
@@ -10,11 +12,17 @@ void main() async {
   await Put().report();
   await PutInTx().report();
   await PutMany().report();
-  await PutAsyncSequential().report();
+  await PutManyAsync().report();
+  await PutAndGetManyAsync().report();
+  await PutQueuedAwaitResult().report();
+  await PutQueued().report();
+  await PutAsync().report();
+  await PutAndGetAsync().report();
+  await PutQueuedAwaitResultParallel().report();
+  await PutQueuedParallel().report();
   await PutAsyncParallel().report();
   await PutAsyncToList().report();
   await PutAsyncAwait().report();
-  await PutQueued().report();
   await RunInTx().report();
   await RunInTxAsync().report();
   await RunAsync().report();
@@ -30,10 +38,12 @@ class Put extends DbBenchmark {
   void runIteration(int i) => box.put(items[i]);
 }
 
-class PutInTx extends DbBenchmark {
-  final items = prepareTestEntities(count, assignedIds: true);
+const putManyCount = 200000;
 
-  PutInTx() : super('$PutInTx', iterations: 1, coefficient: 1 / count);
+class PutInTx extends DbBenchmark {
+  final items = prepareTestEntities(putManyCount, assignedIds: true);
+
+  PutInTx() : super('$PutInTx', iterations: 1, coefficient: 1 / putManyCount);
 
   @override
   void runIteration(int i) =>
@@ -41,31 +51,117 @@ class PutInTx extends DbBenchmark {
 }
 
 class PutMany extends DbBenchmark {
-  final items = prepareTestEntities(count, assignedIds: true);
+  final items = prepareTestEntities(putManyCount, assignedIds: true);
 
-  PutMany() : super('$PutMany', iterations: 1, coefficient: 1 / count);
+  PutMany() : super('$PutMany', iterations: 1, coefficient: 1 / putManyCount);
 
   @override
   void runIteration(int i) => box.putMany(items);
 }
 
-/// Runs putAsync one-by-one, use to measure time of a single call.
-class PutAsyncSequential extends DbBenchmark {
-  static const count = 10;
-  final items = prepareTestEntities(count, assignedIds: true);
+class PutManyAsync extends DbBenchmark {
+  final items = prepareTestEntities(putManyCount, assignedIds: true);
 
-  PutAsyncSequential() : super('$PutAsyncSequential', iterations: count);
+  PutManyAsync()
+      : super('$PutManyAsync', iterations: 1, coefficient: 1 / putManyCount);
+
+  @override
+  void runIteration(int i) => box.putManyAsync(items);
+}
+
+class PutAndGetManyAsync extends DbBenchmark {
+  final items = prepareTestEntities(putManyCount, assignedIds: true);
+
+  PutAndGetManyAsync()
+      : super('$PutAndGetManyAsync',
+            iterations: 1, coefficient: 1 / putManyCount);
+
+  @override
+  void runIteration(int i) => box.putAndGetManyAsync(items);
+}
+
+const putAsyncCount = 10;
+
+/// Runs putAsync one-by-one, use to measure time of a single call.
+class PutQueuedAwaitResult extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncCount, assignedIds: true);
+
+  PutQueuedAwaitResult()
+      : super('$PutQueuedAwaitResult', iterations: putAsyncCount);
+
+  @override
+  FutureOr<void> runIteration(int iteration) =>
+      box.putQueuedAwaitResult(items[iteration]);
+}
+
+/// Runs putQueued one-by-one, use to measure time of a single call.
+class PutQueued extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncCount, assignedIds: true);
+
+  PutQueued() : super('$PutQueued', iterations: putAsyncCount);
+
+  @override
+  FutureOr<void> runIteration(int iteration) {
+    box.putQueued(items[iteration]);
+    store.awaitQueueSubmitted();
+  }
+}
+
+/// Runs putAsync one-by-one, use to measure time of a single call.
+class PutAsync extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncCount, assignedIds: true);
+
+  PutAsync() : super('$PutAsync', iterations: putAsyncCount);
 
   @override
   FutureOr<void> runIteration(int iteration) => box.putAsync(items[iteration]);
 }
 
+/// Runs one-by-one, use to measure time of a single call.
+class PutAndGetAsync extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncCount, assignedIds: true);
+
+  PutAndGetAsync() : super('$PutAndGetAsync', iterations: putAsyncCount);
+
+  @override
+  FutureOr<void> runIteration(int iteration) =>
+      box.putAndGetAsync(items[iteration]);
+}
+
+const putAsyncParallelCount = 100;
+
+/// Runs many PutQueuedAwaitResult calls in parallel and waits until the last one completes,
+/// use to assert parallelization capability.
+class PutQueuedAwaitResultParallel extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncParallelCount, assignedIds: true);
+
+  PutQueuedAwaitResultParallel() : super('$PutQueuedAwaitResultParallel');
+
+  @override
+  Future<void> run() async =>
+      await Future.wait(items.map(box.putQueuedAwaitResult));
+}
+
+/// Runs many putQueuedParallel calls in parallel and waits until the last one
+/// completes, use to assert parallelization capability.
+class PutQueuedParallel extends DbBenchmark {
+  final items = prepareTestEntities(putAsyncParallelCount, assignedIds: true);
+
+  PutQueuedParallel() : super('$PutQueuedParallel', iterations: 1);
+
+  @override
+  Future<void> run() async {
+    items.forEach(box.putQueued);
+    store.awaitQueueSubmitted();
+  }
+}
+
 /// Runs many putAsync calls in parallel and waits until the last one completes,
 /// use to assert parallelization capability.
 class PutAsyncParallel extends DbBenchmark {
-  final items = prepareTestEntities(count, assignedIds: true);
+  final items = prepareTestEntities(putAsyncParallelCount, assignedIds: true);
 
-  PutAsyncParallel() : super('$PutAsyncParallel');
+  PutAsyncParallel() : super('$PutAsyncParallel', iterations: 1);
 
   @override
   Future<void> run() async => await Future.wait(items.map(box.putAsync));
@@ -80,9 +176,9 @@ class PutAsyncToList extends DbBenchmark {
 
   @override
   Future<void> run() async {
-    final futures = items.map(box.putAsync).toList(growable: false);
+    final futures = items.map(box.putQueuedAwaitResult).toList(growable: false);
     await Future.wait(futures);
-    store.awaitAsyncSubmitted();
+    store.awaitQueueSubmitted();
   }
 }
 
@@ -93,20 +189,8 @@ class PutAsyncAwait extends DbBenchmark {
 
   @override
   Future<void> run() async {
-    items.forEach((item) async => await box.putAsync(item));
-    store.awaitAsyncSubmitted();
-  }
-}
-
-class PutQueued extends DbBenchmark {
-  final items = prepareTestEntities(count, assignedIds: true);
-
-  PutQueued() : super('$PutQueued', iterations: count);
-
-  @override
-  Future<void> run() async {
-    items.forEach(box.putQueued);
-    store.awaitAsyncSubmitted();
+    items.forEach((item) async => await box.putQueuedAwaitResult(item));
+    store.awaitQueueSubmitted();
   }
 }