+++ objectbox/lib/src/native/query/query.dart
@@ -671,7 +671,17 @@ class Query<T> {
   Query._(this._store, Pointer<OBX_query_builder> cBuilder, this._entity)
       : _cQuery = checkObxPtr(C.query(cBuilder), 'create query') {
     initializeDartAPI();
+    _attachFinalizer();
+  }
 
+  Query._fromConfiguration(this._store, _QueryConfiguration<T> configuration)
+      : _cQuery = Pointer.fromAddress(configuration.queryAddress),
+        _entity = configuration.entity {
+    initializeDartAPI();
+    _attachFinalizer();
+  }
+
+  void _attachFinalizer() {
     // Keep the finalizer so we can detach it when close() is called manually.
     _cFinalizer =
         C.dartc_attach_finalizer(this, native_query_close, _cQuery.cast(), 256);
@@ -781,6 +791,16 @@ class Query<T> {
     return result;
   }
 
+  // Static callback to avoid over-capturing due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983).
+  static T? _findFirstAsyncCallback<T>(
+          Store store, _QueryConfiguration<T> configuration) =>
+      _asyncCallbackImpl<T, T?>(
+          store, configuration, (query) => query.findFirst());
+
+  /// Like [findFirst], but runs the query operation asynchronously in a worker
+  /// isolate.
+  Future<T?> findFirstAsync() => _findAsyncImpl(_findFirstAsyncCallback<T>);
+
   /// Finds the only object matching the query. Returns null if there are no
   /// results or throws [NonUniqueResultException] if there are multiple objects
   /// matching.
@@ -813,6 +833,16 @@ class Query<T> {
     return result;
   }
 
+  // Static callback to avoid over-capturing due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983).
+  static T? _findUniqueAsyncCallback<T>(
+          Store store, _QueryConfiguration<T> configuration) =>
+      _asyncCallbackImpl<T, T?>(
+          store, configuration, (query) => query.findUnique());
+
+  /// Like [findUnique], but runs the query operation asynchronously in a worker
+  /// isolate.
+  Future<T?> findUniqueAsync() => _findAsyncImpl(_findUniqueAsyncCallback<T>);
+
   /// Finds Objects matching the query and returns their IDs.
   List<int> findIds() {
     final idArrayPtr = checkObxPtr(C.query_find_ids(_ptr), 'find ids');
@@ -826,6 +856,16 @@ class Query<T> {
     }
   }
 
+  // Static callback to avoid over-capturing due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983).
+  static List<int> _findIdsAsyncCallback<T>(
+          Store store, _QueryConfiguration<T> configuration) =>
+      _asyncCallbackImpl<T, List<int>>(
+          store, configuration, (query) => query.findIds());
+
+  /// Like [findIds], but runs the query operation asynchronously in a worker
+  /// isolate.
+  Future<List<int>> findIdsAsync() => _findAsyncImpl(_findIdsAsyncCallback<T>);
+
   /// Finds Objects matching the query.
   List<T> find() {
     final result = <T>[];
@@ -836,6 +876,32 @@ class Query<T> {
     reachabilityFence(this);
     return result;
   }
+
+  // Static callback to avoid over-capturing due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983).
+  static List<T> _findAsyncCallback<T>(
+          Store store, _QueryConfiguration<T> configuration) =>
+      _asyncCallbackImpl<T, List<T>>(
+          store, configuration, (query) => query.find());
+
+  /// Like [find], but runs the query operation asynchronously in a worker
+  /// isolate.
+  Future<List<T>> findAsync() => _findAsyncImpl(_findAsyncCallback<T>);
+
+  // Static callback to avoid over-capturing due to [dart-lang/sdk#36983](https://github.com/dart-lang/sdk/issues/36983).
+  static R _asyncCallbackImpl<T, R>(Store store,
+      _QueryConfiguration<T> configuration, R Function(Query<T>) action) {
+    final query = Query._fromConfiguration(store, configuration);
+    try {
+      return action(query);
+    } finally {
+      query.close();
+    }
+  }
+
+  /// Runs the given query callback on a worker isolate and returns the result.
+  Future<R> _findAsyncImpl<R>(
+          R Function(Store, _QueryConfiguration<T>) callback) =>
+      _store.runAsync(callback, _QueryConfiguration(this));
 
   /// Finds Objects matching the query, streaming them while the query executes.
   ///
@@ -1156,4 +1222,14 @@ class _StreamIsolateMessage {
   final List<int> sizes;
 
   const _StreamIsolateMessage(this.dataPtrAddresses, this.sizes);
+}
+
+class _QueryConfiguration<T> {
+  final int queryAddress;
+  final EntityDefinition<T> entity;
+
+  /// Creates a configuration to send to an isolate by cloning the native query.
+  _QueryConfiguration(Query<T> query)
+      : queryAddress = query._clone().address,
+        entity = query._entity;
 }